refactor: ServeyRepository로 JPARepository, QueryDSL 동시 사용 설계
