+++ src/main/java/com/example/olebackend/domain/mapping/Survey.java
@@ -1,0 +1,28 @@
+package com.example.olebackend.domain.mapping;
+
+import com.example.olebackend.domain.Lesson;
+import com.example.olebackend.domain.Member;
+import com.example.olebackend.domain.common.BaseEntity;
+import lombok.*;
+
+import javax.persistence.*;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class Survey extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id ;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "lesson_id")
+    private Lesson lesson;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id")
+    private Member member;
+}
