+++ src/main/java/com/example/olebackend/repository/specification/SurveyRepositoryCustom.java
@@ -1,0 +1,13 @@
+package com.example.olebackend.repository.specification;
+
+import com.example.olebackend.domain.Lesson;
+import com.example.olebackend.web.dto.SurveyRequest;
+
+import java.util.List;
+
+public interface SurveyRepositoryCustom {
+    public List<Lesson> strictFiltering(SurveyRequest.SurveyCondition condition) ;
+    public List<Lesson> normalFiltering(SurveyRequest.SurveyCondition condition) ;
+    public List<Lesson> lenientFiltering(SurveyRequest.SurveyCondition condition) ;
+
+}
